+++ com.github.fhd.eclipsecolortheme/src/com/github/fhd/eclipsecolortheme/mapper/JavaEditorMapper.java
@@ -31,37 +31,42 @@ public class JavaEditorMapper extends ThemePreferenceMapper {
         preferences.putBoolean(
                 "semanticHighlighting.localVariableDeclaration.enabled", false);
 
-        putPreference("java_default", theme.get("foreground"));
+        putPreference("java_default", entry(theme.get("foreground")));
         putPreference("java_single_line_comment",
-                      theme.get("singleLineComment"));
-        putPreference("java_multi_line_comment", theme.get("multiLineComment"));
-        putPreference("java_comment_task_tag", theme.get("commentTaskTag"));
+                      entry(theme.get("singleLineComment")));
+        putPreference("java_multi_line_comment",
+                      entry(theme.get("multiLineComment")));
+        putPreference("java_comment_task_tag",
+                      entry(theme.get("commentTaskTag")));
         putPreference("sourceHoverBackgroundColor",
-                      theme.get("sourceHoverBackground"));
-        putPreference("semanticHighlighting.number.color", theme.get("number"));
-        putPreference("java_string", theme.get("string"));
-        putPreference("java_bracket", theme.get("bracket"));
-        putPreference("java_operator", theme.get("operator"));
-        putPreference("java_keyword_return", theme.get("keyword"));
-        putPreference("java_keyword", theme.get("keyword"));
-        putPreference("semanticHighlighting.class.color", theme.get("class"));
+                      entry(theme.get("sourceHoverBackground")));
+        putPreference("semanticHighlighting.number.color",
+                      entry(theme.get("number")));
+        putPreference("java_string", entry(theme.get("string")));
+        putPreference("java_bracket", entry(theme.get("bracket")));
+        putPreference("java_operator", entry(theme.get("operator")));
+        putPreference("java_keyword_return", entry(theme.get("keyword")));
+        putPreference("java_keyword", entry(theme.get("keyword")));
+        putPreference("semanticHighlighting.class.color",
+                      entry(theme.get("class")));
         putPreference("semanticHighlighting.interface.color",
-                      theme.get("interface"));
-        putPreference("semanticHighlighting.method.color", theme.get("method"));
+                      entry(theme.get("interface")));
+        putPreference("semanticHighlighting.method.color",
+                      entry(theme.get("method")));
         putPreference("semanticHighlighting.methodDeclarationName.color",
-                      theme.get("methodDeclarationName"));
+                      entry(theme.get("methodDeclarationName")));
         putPreference("semanticHighlighting.annotation.color",
-                      theme.get("annotation"));
-        putPreference("java_doc_default", theme.get("javadoc"));
-        putPreference("java_doc_link", theme.get("javadocLink"));
-        putPreference("java_doc_keyword", theme.get("javadocKeyword"));
-        putPreference("java_doc_tag", theme.get("javadocTag"));
+                      entry(theme.get("annotation")));
+        putPreference("java_doc_default", entry(theme.get("javadoc")));
+        putPreference("java_doc_link", entry(theme.get("javadocLink")));
+        putPreference("java_doc_keyword", entry(theme.get("javadocKeyword")));
+        putPreference("java_doc_tag", entry(theme.get("javadocTag")));
         putPreference("semanticHighlighting.localVariable.color",
-                      theme.get("localVariable"));
+                      entry(theme.get("localVariable")));
         putPreference("semanticHighlighting.localVariableDeclaration.color",
-                      theme.get("localVariableDeclaration"));
+                      entry(theme.get("localVariableDeclaration")));
         putPreference("semanticHighlighting.staticFinalField.color",
-                      theme.get("staticFinalField"));
+                      entry(theme.get("staticFinalField")));
     }
 
     @Override